riveredge-root/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riveredge-root"
version = "1.0.0"
description = "RiverEdge SaaS 多组织框架 - 系统级后端（根）"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]>=0.30.0",
    "pydantic==2.5.0",
    "pydantic-settings>=2.5.0",
    "tortoise-orm==0.20.1",  # 使用稳定版本 0.20.1（与FastAPI 0.104.1完美兼容）
    "aerich==0.7.1",
    "asyncpg==0.29.0",  # 使用稳定版本 0.29.0（Python 3.12 支持预编译包）
    "redis>=5.0.0",  # redis-p, 使用 redis.asyncio 异步接口
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.aerich]
tortoise_orm = "core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./src"
